% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{collapse_repeated_weight}
\alias{collapse_repeated_weight}
\title{Collapse repeated PSU or haul total weights stored on multiple LFD rows.}
\usage{
collapse_repeated_weight(x)
}
\arguments{
\item{x}{numeric-like vector.}
}
\value{
Single numeric value or \code{NA_real_}.
}
\description{
Collapse repeated PSU or haul total weights stored on multiple LFD rows.
}
\author{
stat4RDBFIS Core Team
}
