% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{is_missing_token}
\alias{is_missing_token}
\title{Treat NA, the string "NA", and empty strings as missing.}
\usage{
is_missing_token(x)
}
\arguments{
\item{x}{vector.}
}
\value{
Logical vector.
}
\description{
Treat NA, the string "NA", and empty strings as missing.
}
\author{
stat4RDBFIS Core Team
}
