% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_grc.R
\name{method_grc}
\alias{method_grc}
\title{GRC landings LFD raising method}
\usage{
method_grc(data, land)
}
\arguments{
\item{data}{data.frame. Prepared sampling LFD data for one stock.}

\item{land}{data.frame. Landings totals aggregated by stratum.}
}
\value{
A data.frame with raised numbers-at-length in long format.
}
\description{
Apply the Greece-specific raising method to prepared LFD inputs.
}
\details{
This low-level method is called by \code{landings_rdbfis()}. It raises
numbers-at-length by haul or trip catch weight and then to total landings at
stratum level.
}
\author{
stat4RDBFIS Core Team
}
