% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat4RDBFIS-package.R
\docType{package}
\name{stat4RDBFIS-package}
\alias{stat4RDBFIS}
\alias{stat4RDBFIS-package}
\title{stat4RDBFIS}
\description{
Quality control, data preparation, and statistical raising tools for RCG
landings (CL) and sampling (CS) data from Mediterranean and Black Sea
fisheries.
}
\details{
The package was developed within the RDBFIS III project (Activity 10) by
the stat4RDBFIS Core Team: Vicky Sgardeli (HCMR), Ilaria Costantini
(CNR-IRBIM), Kostas Touloumis (FRI), Alessandro Mannini (CNR-IRBIM), and
Stefanos Kavadas (HCMR).

The main workflow includes landings and sampling quality control,
crosschecks between CL and CS templates, landings and discard LFD raising,
maturity-at-age and maturity-at-length estimation, sex-ratio estimation,
age-length keys, and discard ratio estimation.
}
\author{
stat4RDBFIS Core Team
}
\keyword{internal}
