% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_cc.R
\name{trip_cc}
\alias{trip_cc}
\title{Generic trip and commercial-category LFD raising method}
\usage{
trip_cc(data, land)
}
\arguments{
\item{data}{data.frame. Prepared sampling LFD data for one stock.}

\item{land}{data.frame. Landings or discard totals aggregated by stratum.}
}
\value{
A data.frame with raised numbers-at-length in long format.
}
\description{
Apply a generic raising method by trip and commercial category.
}
\details{
This low-level method is called by \code{landings_rdbfis()} or
\code{discards_rdbfis()}. It raises numbers-at-length to catch weight by trip
and commercial category, then to totals at stratum level.
}
\author{
stat4RDBFIS Core Team
}
