% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_h.R
\name{trip_h}
\alias{trip_h}
\title{Generic haul or trip LFD raising method}
\usage{
trip_h(data, land)
}
\arguments{
\item{data}{data.frame. Prepared sampling LFD data for one stock.}

\item{land}{data.frame. Landings or discard totals aggregated by stratum.}
}
\value{
A data.frame with raised numbers-at-length in long format.
}
\description{
Apply a generic raising method based on haul or trip catch weights.
}
\details{
This low-level method is called by \code{landings_rdbfis()} or
\code{discards_rdbfis()}. It raises numbers-at-length using catch weight and
then to totals at stratum level.
}
\author{
stat4RDBFIS Core Team
}
