% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigneau_mahevas.R
\name{vigneau_mahevas}
\alias{vigneau_mahevas}
\title{Vigneau-Mahevas LFD raising method}
\usage{
vigneau_mahevas(data, land)
}
\arguments{
\item{data}{data.frame. Prepared sampling LFD data for one stock.}

\item{land}{data.frame. Landings or discard totals aggregated by stratum.}
}
\value{
A data.frame with raised numbers-at-length in long format.
}
\description{
Apply the Vigneau-Mahevas raising method to prepared LFD inputs.
}
\details{
This low-level method is called by \code{landings_rdbfis()} or
\code{discards_rdbfis()}. It raises the summed samples to totals at stratum
level.
}
\author{
stat4RDBFIS Core Team
}
